+++ packages/eslint-config/index.js
@@ -17,16 +17,22 @@ module.exports = {
     'airbnb',
     'plugin:@typescript-eslint/recommended',
     'prettier',
-    'prettier/@typescript-eslint',
-    'prettier/react',
     ...localExtends,
   ],
 
-  globals: {
-    // testcafe
-    fixture: false,
-    test: false,
-  },
+  overrides: [
+    {
+      files: ['**/*.js'],
+      rules: {
+        '@typescript-eslint/no-unsafe-assignment': 'off',
+        '@typescript-eslint/no-unsafe-call': 'off',
+        '@typescript-eslint/no-unsafe-member-access': 'off',
+        '@typescript-eslint/no-unsafe-return': 'off',
+        '@typescript-eslint/no-var-requires': 'off',
+        '@typescript-eslint/restrict-template-expressions': 'off',
+      },
+    },
+  ],
 
   plugins: [
     'import',
@@ -41,6 +47,7 @@ module.exports = {
 
   rules: {
     // These off-by-default or configurable rules are good and we like having them on
+    '@typescript-eslint/array-type': 'error',
     eqeqeq: 'error',
     'no-console': 'error',
     'no-eval': 'error',
@@ -70,50 +77,13 @@ module.exports = {
     ],
     'simple-import-sort/imports': 'error',
 
-    // These rules could be useful, but we haven't gotten around to trying them out
-    '@typescript-eslint/array-type': 'off',
-    '@typescript-eslint/ban-types': 'off',
-    '@typescript-eslint/camelcase': 'off',
-    '@typescript-eslint/explicit-member-accessibility': 'off',
-    '@typescript-eslint/explicit-module-boundary-types': 'off',
-    '@typescript-eslint/generic-type-naming': 'off',
-    '@typescript-eslint/member-ordering': 'off',
-    '@typescript-eslint/no-empty-function': 'off',
-    '@typescript-eslint/no-explicit-any': 'off',
-    '@typescript-eslint/no-non-null-assertion': 'off',
-    '@typescript-eslint/no-require-imports': 'off',
-    '@typescript-eslint/no-unnecessary-condition': 'off',
-    '@typescript-eslint/no-use-before-define': 'off',
-    '@typescript-eslint/no-var-requires': 'off',
-    '@typescript-eslint/prefer-includes': 'off',
-    '@typescript-eslint/promise-function-async': 'off',
-    '@typescript-eslint/require-array-sort-compare': 'off',
-    '@typescript-eslint/strict-boolean-expressions': 'off',
-    'no-await-in-loop': 'off',
-    'no-cond-assign': 'off',
-    'no-plusplus': 'off',
-    'no-restricted-globals': 'off',
-    'import/order': 'off',
-    'lines-between-class-members': 'off',
-    'react/no-unused-prop-types': 'off',
-    'consistent-return': 'off',
-    'default-case': 'off',
-    'import/named': 'off',
-    'import/namespace': 'off',
-    'import/no-extraneous-dependencies': 'off',
-    'import/no-unresolved': 'off',
-    'jsx-a11y/label-has-associated-control': 'off',
-    'no-multi-assign': 'off',
-    'no-sparse-arrays': 'off',
-    'no-template-curly-in-string': 'off',
-    'prefer-template': 'off',
-    'react/destructuring-assignment': 'off',
-    'react/display-name': 'off',
-    'react/sort-comp': 'off',
-
     // These functional rules are annoying and we generally don't want them on
+    // Note that these are only the rules we don't want in *any* consuming repository
+    // Rules to be temporarily disabled for convenience should be done so in those repositories
+    '@typescript-eslint/ban-types': 'off',
     '@typescript-eslint/consistent-type-definitions': 'off',
     '@typescript-eslint/explicit-function-return-type': 'off',
+    '@typescript-eslint/explicit-module-boundary-types': 'off',
     '@typescript-eslint/no-empty-interface': 'off',
     '@typescript-eslint/no-magic-numbers': 'off',
     '@typescript-eslint/no-type-alias': 'off',
@@ -122,12 +92,15 @@ module.exports = {
     'arrow-body-style': 'off',
     camelcase: 'off',
     'class-methods-use-this': 'off',
+    'default-case': 'off',
     'func-names': 'off',
     'global-require': 'off',
     'import/export': 'off',
     'import/extensions': 'off',
     'import/no-default-export': 'off',
+    'import/no-unresolved': 'off',
     'import/prefer-default-export': 'off',
+    'lines-between-class-members': 'off',
     'max-classes-per-file': 'off',
     'no-bitwise': 'off',
     'no-case-declarations': 'off',
@@ -140,9 +113,11 @@ module.exports = {
     'no-return-assign': 'off',
     'no-return-await': 'off',
     'no-shadow': 'off',
+    'no-sparse-arrays': 'off',
     'no-underscore-dangle': 'off',
     'no-use-before-define': 'off',
     'no-useless-constructor': 'off',
+    'prefer-template': 'off',
     'react/jsx-filename-extension': 'off',
     'react/jsx-no-bind': 'off',
     'react/jsx-pascal-case': 'off',
@@ -153,32 +128,6 @@ module.exports = {
     'react/static-property-placement': 'off',
     'react/style-prop-object': 'off',
     strict: 'off',
-
-    // These rules have been deprecated in their plugins or overridden but not yet removed from presets
-    'jsx-a11y/label-has-for': 'off',
-    'jsx-a11y/accessible-emoji': 'off',
-    'no-unused-expressions': 'off',
-
-    // These style rules conflict with Prettier but aren't disabled by its plugins
-    '@typescript-eslint/quotes': 'off',
-
-    'no-restricted-imports': [
-      'warn',
-      {
-        paths: [
-          '@codecademy/gamut/dist/deprecated',
-          {
-            name: 'moment',
-            message: 'Please use date-fns instead.',
-          },
-          {
-            name: 'moment-timezone',
-            message: 'Please use date-fns-tz instead.',
-          },
-        ],
-        patterns: ['@codecademy/gamut/dist/deprecated/*'],
-      },
-    ],
   },
 
   settings: {